+++ Backend/CommandPatternLauncher/src/MDC/Gamedata/PlayerType/RangeFighter.cs
@@ -7,7 +7,7 @@ namespace MDC.Gamedata.PlayerType
     {
          public override double _attackPower => 0.5;
          public override double _defensePower => 0;
-         public override double _attackRange => 3;
+         public override int _attackRange => 3;
 
     }
 }
